Move analysisprovider to pipedv1/plugin/analysis/analysisprovider
