+++ pkg/app/pipedv1/plugin/analysis/analysisprovider/log/factory/factory.go
@@ -20,8 +20,8 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/pipe-cd/pipecd/pkg/app/pipedv1/analysisprovider/log"
-	"github.com/pipe-cd/pipecd/pkg/app/pipedv1/analysisprovider/log/stackdriver"
+    "github.com/pipe-cd/pipecd/pkg/app/pipedv1/plugin/analysis/analysisprovider/log"
+    "github.com/pipe-cd/pipecd/pkg/app/pipedv1/plugin/analysis/analysisprovider/log/stackdriver"
 	"github.com/pipe-cd/pipecd/pkg/config"
 	"github.com/pipe-cd/pipecd/pkg/model"
 )
