+++ pkg/app/pipedv1/plugin/analysis/analysisprovider/metrics/datadog/datadog.go
@@ -23,7 +23,7 @@ import (
 	"github.com/DataDog/datadog-api-client-go/api/v1/datadog"
 	"go.uber.org/zap"
 
-	"github.com/pipe-cd/pipecd/pkg/app/pipedv1/analysisprovider/metrics"
+    "github.com/pipe-cd/pipecd/pkg/app/pipedv1/plugin/analysis/analysisprovider/metrics"
 )
 
 const (
